+++ src/app/interfaces/model.interface.ts
@@ -2,6 +2,7 @@ import { KeyValue } from '@angular/common';
 import { IManufacturer } from './manufacturer.interface';
 
 export class IModel {
+    id: number;
     name: string;
     images: string[];
     manufacturer: IManufacturer;