+++ src/app/modules/models-module/all-models/all-models.page.ts
@@ -3,6 +3,7 @@ import { AbstractBackNavigationPage } from 'src/app/utils/abstract-back-navigati
 import { BackNavigationService } from '../../../services/back-navigation.service';
 import { DataFetcherService } from '../../../services/data-fetcher.service';
 import { IModel } from 'src/app/interfaces/model.interface';
+import { ComparisonService } from 'src/app/services/comparison.service';
 
 @Component({
   selector: 'app-all-models',
@@ -16,7 +17,8 @@ export class AllModelsPage extends AbstractBackNavigationPage implements OnInit 
 
   constructor(
     backNavigationService: BackNavigationService,
-    private dataFetcherService: DataFetcherService
+    private dataFetcherService: DataFetcherService,
+    private comparisonsService: ComparisonService
   ) {
     super(backNavigationService, { toHome: true, inRoot: false });
     this.textFilter = '';
@@ -24,7 +26,7 @@ export class AllModelsPage extends AbstractBackNavigationPage implements OnInit 
   }
 
   public ngOnInit() {
-    this.models = this.dataFetcherService.AllModels;
+    this.models = this.order(this.dataFetcherService.AllModels);
   }
 
   public get filteredModels(): IModel[] {
@@ -33,6 +35,16 @@ export class AllModelsPage extends AbstractBackNavigationPage implements OnInit 
       : this.models.filter((model: IModel) =>
           model.name.toLocaleLowerCase().includes(this.textFilter.toLocaleLowerCase())
         );
+  }
+
+  public get disabled(): boolean {
+    return this.comparisonsService.comparison.length === 0;
+  }
+
+  private order(array: IModel[]): IModel[] {
+    return array.sort((model1: IModel, model2: IModel) =>
+      model1.name.toLowerCase().localeCompare(model2.name.toLowerCase())
+    );
   }
 
 }