+++ src/app/modules/models-module/models/models.page.ts
@@ -4,6 +4,7 @@ import { BackNavigationService } from '../../../services/back-navigation.service
 import { DataFetcherService } from 'src/app/services/data-fetcher.service';
 import { IModel } from 'src/app/interfaces/model.interface';
 import { IManufacturer } from 'src/app/interfaces/manufacturer.interface';
+import { ComparisonService } from 'src/app/services/comparison.service';
 
 @Component({
   selector: 'app-models',
@@ -17,7 +18,8 @@ export class ModelsPage extends AbstractBackNavigationPage implements OnInit {
 
   constructor(
     backNavigationService: BackNavigationService,
-    private dataFetchService: DataFetcherService
+    private dataFetchService: DataFetcherService,
+    private comparisonsService: ComparisonService
   ) {
     super(backNavigationService, { toHome: false, inRoot: false });
     this.selectedManufacturer = undefined;
@@ -27,6 +29,10 @@ export class ModelsPage extends AbstractBackNavigationPage implements OnInit {
   public ngOnInit() {
     this.selectedManufacturer = this.dataFetchService.getSelectedManufacturer();
     this.models = this.dataFetchService.getModelsByManufacturer(this.selectedManufacturer);
+  }
+
+  public get disabled(): boolean {
+    return this.comparisonsService.comparison.length === 0;
   }
 
 }